+++ rollup.prebundle.config.mjs
@@ -1,8 +1,13 @@
-import path from 'path'
+import { createRequire } from 'node:module'
+import path from 'node:path'
+import url from 'node:url'
 import alias from '@rollup/plugin-alias'
 import json from '@rollup/plugin-json'
 import esbuild from 'rollup-plugin-esbuild'
-import { compilerOptions } from './tsconfig.json'
+
+const require = createRequire(import.meta.url)
+const { compilerOptions } = require('./tsconfig.json')
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url))
 
 export const prebundleAlias = alias({
   entries: [